gpt-tool/pyproject.toml
[tool.poetry]
name = "performance-engineering-with-gpt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "performance_engineering_with_gpt"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
openai = "^0.27.6"
python-dotenv = "^1.0.0"
paramiko = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
